:recycle: build: 重构类型定义文件结构
