+++ package.json
@@ -20,17 +20,18 @@
     "types"
   ],
   "dependencies": {
-    "cocoascript-types": "*",
-    "sketch-internal-types": "*"
+    "cocoascript-types": "latest",
+    "sketch-internal-types": "latest"
   },
   "devDependencies": {
-    "prettier": "^2.1.1",
-    "typescript": "^4.1.3",
+    "@types/jest": "^27.0.2",
     "commitlint": "^12.0.1",
     "commitlint-config-gitmoji": "^2.2.5",
     "husky": "^4.3.8",
+    "prettier": "^2.1.1",
     "semantic-release": "^18.0.0",
-    "semantic-release-config-gitmoji": "^1.2.4"
+    "semantic-release-config-gitmoji": "^1.2.4",
+    "typescript": "^4.1.3"
   },
   "bugs": {
     "url": "https://github.com/sketch-community/sketch-types/issues"
