+++ test/ui.test.ts
@@ -1,0 +1,7 @@
+import { message } from 'sketch/ui';
+
+describe('ui', () => {
+  it('message', () => {
+    message('hello');
+  });
+});
